Migrated to GPLv3, fuck yeah free software
I think GPLv3 is a good license for my work and thus I am migrating this
project to it. On this server-side software it should not have a lot of
impacts since one can modify this and run it without releasing changes,
only if they distribute it.

There is always the possibility to contact me for licensing questions
about this.